backend/requirements.txt
# Django E-Commerce Automation Backend Dependencies
#
# To install all dependencies, run:
#   pip install -r requirements.txt
#
# Note: Make sure you're in your virtual environment before installing.

# Core Django
Django>=5.2.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
mysqlclient>=2.2.0

# Background Tasks
django-q2>=1.9.0

# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# CORS Support
# IMPORTANT: Run 'pip install django-cors-headers' in your virtualenv
django-cors-headers>=4.3.0

# Environment Variables
python-dotenv>=1.0.0

